pyproject.toml
[project]
name = "speak-as-tolkien-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.13"
dependencies = [
    "coqpit>=0.0.17",
    "fastapi>=0.115.12",
    "numpy>=1.24.3",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "torch==2.2.2",
    "torchaudio>=2.2.2",
    "webrtcvad>=2.0.10",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "ipdb>=0.13.13",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "jupyterlab>=4.4.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.2",
    "pytest>=8.3.5",
]
